- zu analysierende Audio-Daten über AudioSample und nicht mehr Audio-Player

+++ src/main/java/ch/fhnw/meco/processor/IVideoProcessor.java
@@ -4,7 +4,7 @@ import java.awt.image.BufferedImage;
 
 public interface IVideoProcessor {
 
-    public void processAudio(byte[] byteArray);
+    public void processAudio(float[] byteArray);
 
     public BufferedImage processImage(BufferedImage image);
 }
